Organize components into folders and change Import
