+++ src/places/components/PlaceItem/PlaceItem.css
