+++ src/user/components/UserItem/UserItem.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import Avatar from "../../shared/components/UIElements/Avatar/Avatar";
-import Card from "../../shared/components/UIElements/Card/Card";
+import Avatar from "../../../shared/components/UIElements/Avatar/Avatar";
+import Card from "../../../shared/components/UIElements/Card/Card";
 import "./UserItem.css";
 
 const UserItem = (props) => {
